refactor(enums): move status enums to dedicated package
